new `safe-wasm` JS dist to fix `unsafe-eval-wasm` header req

+++ bundle.js
@@ -3,8 +3,13 @@ import fs from 'fs';
 
 const embed = (file) => zlib.gzipSync(fs.readFileSync(file), { level: 9 }).toString('base64').replace(/\//g,'$').replace(/O/g,'#');
 
+// Safe Wasm version does not include the Wasm binary in the bundle
+const args = process.argv.slice(2);
+const isSafeWasmVersion = args?.includes('--safe-wasm');
+
 const version = process.env.npm_package_version;
-const outFile = `./public/dist/micrio.min.js`;
+const outFile = `./public/dist/micrio.min${isSafeWasmVersion ? '.safe-wasm' : ''}.js`;
+const outWasmFile = `./public/dist/micrio.min.wasm`;
 const buildDir = './public/build/';
 
 const files=  {
@@ -23,12 +28,18 @@ matches.filter((s,i) => matches.indexOf(s)==i).forEach(sel => {
 fs.writeFileSync(files.css, file);
 
 // Wasm params
-file = `const [_c,_u,_b]=arguments,_css=document.createElement('style');
-_css.className='micrio-interface';_css.textContent=_c;
-document.head.insertBefore(_css,document.head.firstChild);
-`+fs.readFileSync(files.js).toString()
-	.replace('{b64:"EXTERNAL_WASM"}','{b64:_b,ugz:_u}');
-fs.writeFileSync(files.js, file);
+if(isSafeWasmVersion) {
+	file = `const [_c,_u,_b]=arguments,_css=document.createElement('style');
+	_css.className='micrio-interface';_css.textContent=_c;
+	document.head.insertBefore(_css,document.head.firstChild);
+	`+fs.readFileSync(files.js);
+	fs.writeFileSync(files.js, file);
+}
+else {
+	file = fs.readFileSync(files.js).toString()
+		.replace('{b64:"EXTERNAL_WASM"}','{b64:_b,ugz:_u}');
+	fs.writeFileSync(files.js, file);
+}
 
 fs.writeFileSync(outFile, Buffer.concat([
 	Buffer.from([
@@ -39,6 +50,9 @@ fs.writeFileSync(outFile, Buffer.concat([
 	Buffer.from([
 		"((M,I,C,R='O')=>{\n",
 			"if(!(typeof window !== undefined && typeof window !== 'undefined')) return;",
+			...(isSafeWasmVersion ? [
+				`M='',C=null,Promise.resolve().then(()=>`
+			] : [
 			`M='${embed(files.wasm)}',`,
 			"I=window.DecompressionStream,",
 			"C=(b,t,d,a)=>",
@@ -46,32 +60,41 @@ fs.writeFileSync(outFile, Buffer.concat([
 				"d=I?new Response(new Blob([new Uint8Array(a)]).stream().pipeThrough(new I('gzip'))):(new Zlib.Gunzip(a)).decompress())",
 				"&&(I?t?d.arrayBuffer():d.text():t?d:new self.TextDecoder('utf-8').decode(d));",
 			"(I?new Promise(a=>a()):fetch('https://r2.micr.io/gunzip.min.js').then(r=>r.text()).then(js=>Function(js)())).then(()=>",
+			]),
 			`(function(){${fs.readFileSync(files.js)}})(\`${fs.readFileSync(files.css)}\`,C,M))`,
 		"})();",
 	].join(''))
 ]));
 
-console.log();
+if(isSafeWasmVersion) console.log();
+else {
+	fs.copyFileSync(files.wasm, outWasmFile);
+	console.info('\x1b[36m%s\x1b[0m', `created ${outWasmFile}`);
+}
+
 console.info('\x1b[36m%s\x1b[0m', `created ${outFile}`);
 
-const dFile = outFile.replace('.js','.d.ts');
-fs.writeFileSync(dFile, Buffer.concat([
-	Buffer.from([
-		"declare module '@micrio/client' {",
-		"\timport type { Readable, Writable } from 'svelte/store';",
-		...fs.readFileSync('./out.d.ts').toString().replace(/    /mg,'\t').split('\n').filter(l => /^\s/.test(l) && !/^\s*import/.test(l)),
-		"}"
-	].join('\n')),
-	fs.readFileSync('./docs/store.d.ts.txt')
-]));
-fs.rmSync('./out.d.ts');
-fs.rmSync(files.css);
-fs.rmSync(files.js);
+// Final, full version also finalizes the bundling process
+if(!isSafeWasmVersion) {
+	const dFile = outFile.replace('.js','.d.ts');
+	fs.writeFileSync(dFile, Buffer.concat([
+		Buffer.from([
+			"declare module '@micrio/client' {",
+			"\timport type { Readable, Writable } from 'svelte/store';",
+			...fs.readFileSync('./out.d.ts').toString().replace(/    /mg,'\t').split('\n').filter(l => /^\s/.test(l) && !/^\s*import/.test(l)),
+			"}"
+		].join('\n')),
+		fs.readFileSync('./docs/store.d.ts.txt')
+	]));
+	fs.rmSync('./out.d.ts');
+	fs.rmSync(files.css);
+	fs.rmSync(files.js);
 
-fs.writeFileSync('./src/ts/version.ts', `/**
+	fs.writeFileSync('./src/ts/version.ts', `/**
  * Defines the current version of the Micrio library.
  * This constant is used internally and exposed statically via \`HTMLMicrioElement.VERSION\`.
  */
 export const VERSION = '${version}';\n`);
 
-console.info('\x1b[36m%s\x1b[0m', `created ${dFile}`);
+	console.info('\x1b[36m%s\x1b[0m', `created ${dFile}`);
+}
